+++ src/info.ts
@@ -168,13 +168,10 @@ export class Info {
 
     /**
      * Return the set of available features in this environment.
-     * Some features of Luxon are not available in all environments. For example, on older browsers, timezone support is not available. Use this function to figure out if that's the case.
+     * Some features of Luxon are not available in all environments. For example, on older browsers, relative time formatting support is not available. Use this function to figure out if that's the case.
      * Keys:
-     * * `zones`: whether this environment supports IANA timezones
-     * * `intlTokens`: whether this environment supports internationalized token-based formatting/parsing
-     * * `intl`: whether this environment supports general internationalization
      * * `relative`: whether this environment supports relative time formatting
-     * @example Info.features() //=> { intl: true, intlTokens: false, zones: true, relative: false }
+     * @example Info.features() //=> { relative: false }
      * @return {Object}
      */
     static features(): Features {